Ellipsize rendered call data above 1 kilobyte
Byeond a certain point, it is pointless to print the entire hex
call data because it is practically impossible to verify through
visual means. In this case, it is taking up valuable screen space
and displacing potentially important information elsewhere. I feel
1 kilobyte is the point at which data takes up too much space to be
useful.

If a user nevertheless needs to see the call data without ellipsis,
this can be achieved by passing the --verbose option and checking the
log messages.