Reduce default data truncation limit to 500 bytes
In day-to-day usage, I have found that any bytes data above 500 bytes
in the panels is just noise on the screen, as it is just being ignored.
It is also impractical to verify at that size. As vertical screen
space is valuable, it seems like a good idea to further reduce the data
truncation limit in order to eliminate unnecessary information from
the interface.

Note that panel data can still be viewed without truncation using the
--expand option.