Ensure unique signatures in preview & exec
Even though duplicate valid signatures from the same owner are
disallowed in the Safe contract, they were permitted in the `preview`
and `exec` commands. So a transaction with duplicate valid signatures
would count towards the threshold, and could therefore end up showing
as executable when in fact is not. Avoid this scenario by ensuring all
signatures are unique.